refactor(emotion): 提取 applyEmotionEvalRaw 共用解析+落 buff (为情绪评估上 worker 铺路)
把 evaluateEmotionBackground 内联的「解析 eval raw JSON → sanitize buffs → DB.saveCharacter →
广播 emotion-updated → 返回 innerState」抽到 utils/emotionApply.ts:applyEmotionEvalRaw,
本地路径继续用. 下一步 instant 路径让 worker 跑 eval 后推 emotion_update, 客户端 flush 时
复用同一份 apply. 行为不变; 顺手修了 sanitizeBuffs 对 exactOptionalPropertyTypes 的 2 个预存类型错误.
